Fix various character set bugs
Fix: update char_set macros

Previously, the create_character_set macro always created a set with the
name char_set. Added a macro argument to allow the character set to be
given a different name.

Fix: get_char_index macro bug

The call to bin_search in the get_char_index macro did not pass any
arguments for the left and right parameters.

The solution was to pass 0 and CHARSETLEN-1 to the macro, representing
the first and last characters of the char_set.

Fix: missing character char_set bug

Previously the create_char_set function skipped the blank character when
filling the character set.

The solution was to adjust the indexing of the for loop which filled the
the array.