Replace encrypt with rotate
Replaced the encrypt functions and their files with the equivalent
rotate counterpart and implemented rotate_char.

The rotate functions contain an additional parameter, align. Align is
the direction which the input should be rotated, allowing the one set of
functions to be used to encrypt and decrypt input.

Additionally, tied together the main function, and added comments to
the character set and binary search header files.